feat: add configurable options
It's now completely up to the user of the extension whether he wants po files to be generated or not.
Also it's now possible to enable the master file lock by default on boot.
And finally, it's now possible to manually trigger json file generation by clicking on the CODE taskbar item.