feat: rewrite logic for running background tasks
The extension now correctly registers multiple same commands being executed in rapid succession. In such cases, only the last instegation will be called.
Next to that, any running command will be executed when a new same exact command with comes in, because the outputs will never be idempotent, there is no need for the first execution to run to completion.