docs: Update sequence diagram event notation
This change updates the sequence diagram in `sequence.puml` to use
correct UML syntax for events during the activation and deactivation
phases. The previous notation `<<:zap:>>` was incorrect and has been
replaced with `<<event>>` to accurately represent the invocation of
`activate()` and `deactivate()` functions in a way that the github action will be able to process correctly.