feat(configuration): Upgrade config version and features
Upgrade configuration version in `configurationManager` and
`package.json` to ensure consistency across the application. This
change includes an update from `0.0.3` to `0.0.4` in the
configuration version to match the new features introduced.

Additionally, introduce new configuration options for auto-generating
missing translations and setting a DeepL API key. These features aim
to enhance the translation file watcher extension by automating the
generation of missing translation entries and allowing for DeepL
translation integration directly through the extension settings.

This upgrade and feature addition are critical for improving user
experience and extending the functionality of the translation file
watcher, making it more versatile and user-friendly.